requirements.txt
requests
beautifulsoup4
psycopg2-binary
schedule
python-dotenv
selenium